Rename some files and variable zSettings
