+++ src/pages/settings/SettingsComponent.tsx
@@ -2,8 +2,8 @@ import Color from "color";
 import { useEffect, useState } from "react";
 import browser from "webextension-polyfill";
 import { hasMatchingKeys } from "../../common/hasMatchingKeys";
-import { settings } from "../../common/settings/settingsNew";
-import { type Settings } from "../../common/settings/zSettings";
+import { settings } from "../../common/settings/settings";
+import { type Settings } from "../../common/settings/settingsSchema";
 import { Alert } from "./Alert";
 import { CustomHintsSetting } from "./CustomHintsSetting";
 import { ExcludeKeysSetting } from "./ExcludeKeysSetting";
